Add exported database connection
