+++ backend/routes/events.js
@@ -1,6 +1,7 @@
 //Cazandra Jae Lapig
 const express = require("express")
 const router = express.Router()
+const db = require('../database').db;
 
 // Cariel Joyce Maga
 router.get("/", (req, res) => {
