+++ backend/routes/jobpost.js
@@ -1,6 +1,7 @@
 //Cazandra Jae Lapig
 const express = require("express")
 const router = express.Router()
+const db = require('../database').db;
 
 // Cazandra Jae Lapig
 router.get("/", (req, res) => {
