+++ backend/server.js
@@ -8,21 +8,6 @@ const mysql = require('mysql');
 //const bodyParser = require('body-parser');
 
 //Cazandra Jae Lapig
-// MySQL Connection
-const db = mysql.createConnection({
-    host: 'localhost',
-    user: 'root',
-    password: '',
-    database: 'alumaniadb'
-});
-// Connect to MySQL
-db.connect((err) => {
-    if (err) {
-        console.error('Error connecting to MySQL: ' + err.stack);
-        return;
-    }
-    console.log('Connected to MySQL as ID ' + db.threadId);
-});
 
 // Route Imports
 const eventRouter = require("./routes/events");
